test: share the RESP command-frame reader across the mock valkey servers
Two scripted-server tests carried byte-identical copies of the `*N` command
frame reader. Lift it into test-utils as readRespCommands.
